Update on "multiply_integers"
[ghstack-poisoned]
