+++ c10/core/SymInt.h
@@ -6,6 +6,7 @@
 #include <c10/util/intrusive_ptr.h>
 
 #include <memory>
+#include <numeric>
 
 namespace c10 {
 
@@ -184,6 +185,20 @@ class C10_API SymInt {
 };
 
 #undef SKIP_IS_SYMBOLIC_ON_MOBILE
+
+/// Sum of a list of SymInt; accumulates into the c10::SymInt expression
+template <
+    typename C,
+    typename std::enable_if<
+        std::is_same<typename C::value_type, c10::SymInt>::value,
+        int>::type = 0>
+inline c10::SymInt multiply_integers(const C& container) {
+  return std::accumulate(
+      container.begin(),
+      container.end(),
+      c10::SymInt(1),
+      [](c10::SymInt a, c10::SymInt b) { return a * b; });
+}
 
 C10_API std::ostream& operator<<(std::ostream& os, SymInt s);
 } // namespace c10
